Move JEC sets into own files to separate names from implementation
